Homogenized code for joints
